Updating with the new futures examples. Also changing "samples-files" directory on dartlang to "f".
