+++ futures/README.txt
@@ -1,0 +1,5 @@
+The source code for most of these examples is under /web 
+directories but could just as easily be under bin/ directories.
+In fact, if these examples were re-written to use dart:io
+for reading files, they would have to be placed under bin/.
+
