+++ futures/async-await-catch-error/pubspec.yaml
@@ -1,0 +1,7 @@
+name: async_await_catch_error
+description: A sample application that reads a file on dartlang using async/await, and prints the contents; errors are caught.
+
+homepage: https://www.dartlang.org/docs/tutorials/futures/
+
+gist: https://gist.github.com/Sfshaza/847e8d9a356f025ad5d6
+dartpad: https://dartpad.dartlang.org/847e8d9a356f025ad5d6
