+++ futures/futures-api-catch-error/pubspec.yaml
@@ -1,0 +1,7 @@
+name: futures_api_catch_error
+description: A sample application that reads a file on dartlang using the Futures API, and prints the contents; errors are caught.
+
+homepage: https://www.dartlang.org/docs/tutorials/futures/
+
+gist: https://gist.github.com/Sfshaza/a9d90ac89aded272d1ca
+dartpad: https://dartpad.dartlang.org/a9d90ac89aded272d1ca
