--- futures/pubspec.yaml
@@ -1,6 +1,0 @@
-name: futures
-description: A sample application
-
-homepage: https://www.dartlang.org/docs/tutorials/futures/
-gist: https://gist.github.com/Sfshaza/3566574c1a665896dc64
-dartpad: https://dartpad.dartlang.org/3566574c1a665896dc64
