--- futures/web/readme
@@ -1,2 +1,0 @@
-This is a fake readme file to thwart generation of the gist files.
-Please ignore.
